chore: move profile component to core subdomain
